+++ heron/instance/src/python/network/__init__.py
@@ -1,0 +1,4 @@
+'''module for network component for python instance'''
+__all__ = ['metricsmgr_client']
+
+from .metricsmgr_client import MetricsManagerClient
