+++ heron/instance/tests/python/basics/bolt_unittest.py
@@ -12,22 +12,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import unittest
-
-import heron.instance.tests.python.mock_generator as mock_generator
-
-class BoltTest(unittest.TestCase):
-  def setUp(self):
-    pass
-
-  def test_simple_read_and_emit(self):
-    # in_stream == out_stream in this case
-    # send and receive a single primitive tuple
-    bolt = mock_generator.MockBolt()
-    bolt.emit(mock_generator.prim_list)
-    bolt.output_helper.send_out_tuples()
-    bolt._read_tuples_and_execute()
-    self.assertIsNotNone(bolt.received_data_tuple)
+#import unittest
+#
+#import heron.instance.tests.python.mock_generator as mock_generator
+#
+#class BoltTest(unittest.TestCase):
+#  def setUp(self):
+#    pass
+#
+#  def test_simple_read_and_emit(self):
+#    # in_stream == out_stream in this case
+#    # send and receive a single primitive tuple
+#    bolt = mock_generator.MockBolt()
+#    bolt.emit(mock_generator.prim_list)
+#    bolt.output_helper.send_out_tuples()
+#    bolt._read_tuples_and_execute()
+#    self.assertIsNotNone(bolt.received_data_tuple)
